# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Logoviking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface StatusResult {
  ok: boolean;
  message?: string;
}

export interface StringResult {
  value: string;
}

export interface StringListResult {
  items: string[];
}

export interface YoutubeEarningsInput {
  views: number;
  cpm?: number;
  engagementRate?: number;
}

export interface TiktokEarningsInput {
  views: number;
  followers?: number;
}

export interface InstagramEarningsInput {
  followers: number;
  engagementRate?: number;
}

export type EarningsResultBreakdownItem = {
  label: string;
  value: string;
};

export interface EarningsResult {
  low: number;
  avg: number;
  high: number;
  currency: string;
  breakdown?: EarningsResultBreakdownItem[];
}

export interface EngagementRateInput {
  followers: number;
  likes: number;
  comments?: number;
  shares?: number;
}

export interface EngagementRateResult {
  rate: number;
  rating: string;
  tip?: string;
}

export interface ThumbnailDownloadInput {
  url: string;
}

export type ThumbnailDownloadResultThumbnailsItem = {
  quality: string;
  url: string;
};

export interface ThumbnailDownloadResult {
  videoId: string;
  thumbnails: ThumbnailDownloadResultThumbnailsItem[];
}

export interface KeywordInput {
  keyword: string;
}

export interface HashtagInput {
  keyword: string;
  platform?: string;
}

export interface CaptionInput {
  keyword: string;
  tone?: string;
}

export type AiThumbnailResultLayoutsItem = {
  title: string;
  text: string;
  colors: string[];
  tip: string;
};

export interface AiThumbnailResult {
  layouts: AiThumbnailResultLayoutsItem[];
}

export interface MetaTagInput {
  title: string;
  description: string;
  url?: string;
  image?: string;
  keywords?: string;
}

export interface MetaTagResult {
  html: string;
}

export interface KeywordDensityInput {
  text: string;
}

export type KeywordDensityResultKeywordsItem = {
  word: string;
  count: number;
  density: number;
};

export interface KeywordDensityResult {
  wordCount: number;
  keywords: KeywordDensityResultKeywordsItem[];
}

export interface SitemapInput {
  urls: string[];
}

export interface RobotsInput {
  allowAll: boolean;
  sitemapUrl?: string;
  disallowed?: string[];
}

export type SocialSpecsResultPlatformsItemSpecsItem = {
  name: string;
  width: number;
  height: number;
};

export type SocialSpecsResultPlatformsItem = {
  platform: string;
  specs: SocialSpecsResultPlatformsItemSpecsItem[];
};

export interface SocialSpecsResult {
  platforms: SocialSpecsResultPlatformsItem[];
}

export interface CreatorKitResult {
  thumbnailIdea: string;
  title: string;
  hashtags: string[];
  caption: string;
}

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  plan: string;
}

export interface AuthResult {
  ok: boolean;
  user?: User;
}

export interface UserResult {
  user?: User | null;
}

export interface HistoryItem {
  id: number;
  tool: string;
  createdAt: string;
  summary: string;
}

export interface HistoryListResult {
  items: HistoryItem[];
}

export interface BlogPostMeta {
  slug: string;
  title: string;
  excerpt: string;
  category: string;
  readingTime: number;
}

export interface BlogListResult {
  items: BlogPostMeta[];
}

export interface BlogPost {
  slug: string;
  title: string;
  excerpt: string;
  category: string;
  readingTime: number;
  contentHtml: string;
}

export interface ContactInput {
  name: string;
  email: string;
  subject?: string;
  message: string;
}