# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Logoviking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
...
components:
  schemas:
...
    RegisterInput:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
      required: [name, email, password]
    LoginInput:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
      required: [email, password]
    AuthResult:
      type: object
      properties:
        ok: { type: boolean }
        user:
          $ref: "#/components/schemas/User"
      required: [ok]
    User:
      type: object
      properties:
        id: { type: number }
        name: { type: string }
        email: { type: string }
        plan: { type: string }
      required: [id, name, email, plan]
...